Fixes for windows build
These are the several fixes needed to bring the Windows build back up to a working state
after over a year of neglect.

Changes necessary for functionality:

- Upgrade the VS solution and project files from VS2005 to VS2008. We no longer
support compiling with MS Visual Studio 2005.

- Update the VS solution and project files to reflect the source directory restructuring.

- Remove our own definitions of struct pollfd and POLLIN, as these are now defined
in Winsock2.h

- Move an mrn_dbg statement printing a stream_id BEFORE the corresponding stream
is erased. (src/Network.C close_Streams())

- Remove include of config.h from several files.

- Conform lightweight c code to the Windows compiler that only implements parts
of C99.

- Conditionally free the filename string in mrn_printf, since Windows implementation
does not malloc

- Define inline -> __inline, strtok_r -> strtok_s, and strdup -> _strdup

- Declare SIZEOF_{CHAR,FLOAT,DOUBLE} for windows instead of using config.h

- Add old code back into bindPort for Windows that starts at the arbitrary port
49152 and loops until finding a free port. Windows will not automatically assign
a port as Linux does.

- Remove our own definition of size_t, since this is now defined by Windows in VS2008

- Add ssize_t define for lightweight, this still is NOT defined by Windows in VS2008

- Remove assumption from NCSend (Windows-specific) that the first send will always work
and always send the entire message (regular and lightweight).

- Change XPlat library's assumption that a Windows thread handle can be used in place
of a unique thread id across many systems.

- Fix incorrect assumption of lightweight's NCSend (Windows-only) that only one buffer
is ever sent.

- Add XPlat_NCsend and XPlat_NCrecv functions.

- Modify pdr_packet_{data,header} to return int instead of bool, since C code calls
these functions and C does not have a bool. In Windows release mode, the C++ bool
representation was != 0  or 1.

- Add SocketUtils to lightweight to support cross-platform closing of sockets

Some less important changes:

- Remove many warnings in the Windows build

- Enable compiling in Release mode in Windows, not just Debug mode

- Add PRIszt and PRIsszt macros to support cross-platform printing of size_t/ssize_t

+++ tests/test_NativeFilters_lightweight.h
@@ -18,7 +18,7 @@ const int32_t INT32VAL=-17;
 const uint32_t UINT32VAL=17;
 const int64_t INT64VAL=-17;
 const uint64_t UINT64VAL=17;
-const float FLOATVAL=123.456;
+const float FLOATVAL=(const float)123.456;
 const double DOUBLEVAL=123.456;
 
 #endif /* test_nativefilters_lightweight_h */
